passage en abstract des classe autre que Transformateur1
