+++ src/abstraction/eq3Transformateur1/ActeurEQ3.java
@@ -11,7 +11,7 @@ import abstraction.fourni.IActeur;
 import abstraction.fourni.Journal;
 import abstraction.fourni.Variable;
 
-public class ActeurEQ3 implements IActeur{
+public abstract class ActeurEQ3 implements IActeur{
 	protected Variable stockFeves;
 	protected Variable stockChocolat;
 	protected Integer cryptogramme;
@@ -36,10 +36,7 @@ public class ActeurEQ3 implements IActeur{
 				+ "L'équipe se garde la liberté d'acheter de la pâte de moyenne gamme";
 		
 	}
-	
-	public Color getColor() {
-		return new Color(52, 152, 219);
-	}
+
 
 	public void initialiser() {
 	}
@@ -47,8 +44,7 @@ public class ActeurEQ3 implements IActeur{
 	public void setCryptogramme(Integer crypto) {
 		this.cryptogramme = crypto;
 	}
-	public void next() {
-	}
+	
 
 	public List<String> getNomsFilieresProposees() {
 		ArrayList<String> filieres = new ArrayList<String>();
@@ -65,12 +61,7 @@ public class ActeurEQ3 implements IActeur{
 		}
 	}
 	
-	public List<Variable> getIndicateurs() {
-		List<Variable> res=new ArrayList<Variable>();
-		res.add(this.stockFeves);
-		res.add(stockChocolat);
-		return res;
-	}
+	
 
 	public List<Variable> getParametres() {
 		List<Variable> res=new ArrayList<Variable>();
